Adjust FetchTransInfoFromUndo to treat epoch as uint32.

+++ src/include/utils/ztqual.h
@@ -34,7 +34,7 @@ typedef struct ZHeapTupleTransInfo
 #define ZHeapTupleSatisfiesVisibility(tuple, snapshot, buffer, ctid) \
 	((*(snapshot)->zsatisfies) (tuple, snapshot, buffer, ctid))
 
-extern void FetchTransInfoFromUndo(ZHeapTuple undo_tup, uint64 *epoch,
+extern void FetchTransInfoFromUndo(ZHeapTuple undo_tup, uint32 *epoch,
 					   TransactionId *xid, CommandId *cid,
 					   UndoRecPtr *urec_ptr, bool skip_lockers);
 extern void ZHeapTupleGetTransInfo(ZHeapTuple zhtup, Buffer buf,
